change messageid type to long

+++ src/Telegram.Bot/Requests/Getting Updates/GetUpdatesRequest.cs
@@ -17,19 +17,19 @@ namespace Telegram.Bot.Requests
         /// Identifier of the first update to be returned
         /// </summary>
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public int Offset { get; set; }
+        public long Offset { get; set; }
 
         /// <summary>
         /// Limits the number of updates to be retrieved. Values between 1—100 are accepted. Defaults to 100.
         /// </summary>
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public int Limit { get; set; }
+        public long Limit { get; set; }
 
         /// <summary>
         /// Timeout in seconds for long polling. Defaults to 0, i.e. usual short polling.
         /// </summary>
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public int Timeout { get; set; }
+        public long Timeout { get; set; }
 
         /// <summary>
         /// List the types of updates you want your bot to receive. Specify an empty list to receive all updates regardless of type (default). If not specified, the previous setting will be used.
